requirements.txt
# Core Deep Learning and Computer Vision
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Data Processing and Analysis
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web Application
streamlit>=1.28.0
flask>=3.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Optional: For Jupyter notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0
